iwlegacy: Rename label in il_eeprom_init()
Rename a jump label according to the current Linux coding style convention.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>
[Rewrote commit title]
Signed-off-by: Julian Calaby <julian.calaby@gmail.com>

---

This was controversial when introduced, however the change is obvious and
harmless and is, in the worst case, just churn.

I'm only including this as it meets my criteria for sheparding pending
patches: it's sane, obviously correct or reviewable by me, and doesn't
require any work to apply.

This does meet those requirements, however given how controviersial it was
when introduced, I'm not expecting people to be enthusiastic about
applying it.

Thanks,

Julian Calaby