rsi: Move variable initialisation into error code
In rsi_send_data_pkt(), it's a little more logical to assign 'status' in
the actual error handling code as opposed to at the top of the functon.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>
[Deleted controversial bits, rewrote commit message]
Signed-off-by: Julian Calaby <julian.calaby@gmail.com>

---

I'm not fussed if this one goes in or not as the central concept of the
patch can be argued both ways.

Markus had originally added changes to move the setting of adapter down to
just before it was used, however people rightly objected to it, so it's
been removed from this re-send and the commit message re-written to
reflect this.

Thanks,

Julian Calaby