rtl818x_pci: Free invalid rx ring allocations
This hardware cannot deal with ring buffer addresses which have a
non-zero least significant byte. The hardware is likely to do bad
things like overwriting adjacent memory in this case.

That said, no hardware should be producing DMA addresses that don't
fit that criteria, so this code should never be run.

Also add a WARN() telling people their DMA allocator is broken.

Suggested-by: Andrea Merello <andrea.merello@gmail.com>
Signed-off-by: Julian Calaby <julian.calaby@gmail.com>

---

This is the last "obvious" leak in this function, however it's still
not cleaning up appropriately when things fail, in particular it
doesn't unmap everything mapped by pci_map_single() on error and
probably other stuff I don't understand.

I don't have hardware to test this with, so this is only compile
tested. Could someone else please continue on from where I left off?

In particular, the following needs to be done:
1. Address any other leaks in this function
2. Replicate this work across all of the Realtek drivers
3. Replicate the follow-up patches across all of the Realtek drivers

I lack the hardware so I cannot actually test this (also, I don't
know if it's possible to run this code.) however it does compile
clean and the only checkpatch warning is for the long string.

Thanks,

Julian Calaby