rtl818x_pci: Free invalid tx ring allocations
This hardware cannot deal with ring buffer addresses which have a
non-zero least significant byte. The hardware is likely to do bad
things like overwriting adjacent memory in this case.

That said, no hardware should be producing DMA addresses that don't
fit that criteria, so this code should never be run.

Also add a WARN() telling people their DMA allocator is broken.

Signed-off-by: Julian Calaby <julian.calaby@gmail.com>

---

This is an identical fix to the rx ring buffer stuff previously.

Again, I lack the hardware so this is only compile tested.

Thanks,

Julian Calaby