rtl818x_pci: Correct the consistent dma mask
This hardware cannot deal with ring buffer addresses which have a
non-zero least significant byte. The hardware is likely to do bad
things like overwriting adjacent memory in this case.

The ring buffers are allocated consistent, so this should instruct
the DMA allocator to only return addresses we can use.

Suggested-by: Andrea Merello <andrea.merello@gmail.com>
Signed-off-by: Julian Calaby <julian.calaby@gmail.com>

---

This looks obviously correct, however DMA is a subtle beast that I
don't really understand, so this might be wrong.

Also, should this be replicated for the non-coherent mask?

Again, I lack the hardware so this is only compile tested.

Thanks,

Julian Calaby